+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes will be documented in this file.
 
+## Unreleased
+
+### Changes
+
+- **Breaking change**: The renderer now requires Node 0.14 or higher. This is because the renderer uses stateless components internally.
+- **Breaking change**: `allowNode` now receives different properties in the options argument. See `README.md` for more details.
+- **Breaking change**: CommonMark has changed some type names. `Html` is now `HtmlInline`, `Header` is now `Heading` and `HorizontalRule` is now `ThematicBreak`. This affects the `allowedTypes` and `disallowedTypes` options.
+- **Breaking change**: A bug in the `allowedTypes`/`disallowedTypes` and `allowNode` options made them only applicable to certain types. In this version, all types are filtered, as expected.
+
+### Added
+
+- New `renderers` option allows you to customize which React component should be used for rendering given types. See `README.md` for more details. (Espen Hovlandsdal / Guillaume Plique)
+- New `unwrapDisallowed` option allows you to select if the contents of a disallowed node should be "unwrapped" (placed into the disallowed node position). For instance, setting this option to true and disallowing a link would still render the text of the link, instead of the whole link node and all it's children disappearing. (Espen Hovlandsdal)
+
 ## [2.2.2] - 2016-01-22
 
 ### Added